Update app.py
Major backend update.

Replaced lists with a words Python dictionary. It prevent the insertion of duplicate words.

Data structure architecture is now more coherent:
Instead of mapping kanji to kanji and radical information, we create 3 maps. One maps radical to radical information. The second maps kanji to kanji information and unique radical. The third maps words to word information.
So, we can use a word's kanji characters to select a kanji in the kanji map and use a kanji's unique radical identifier to select a radical in the radical map. And we use all maps to fill the SQLite database.